requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
prometheus_client==0.19.0
httpx==0.25.1
pyyaml==6.0.1
numpy
Faker
